MAINT: PR 12906 revisions
* remove the extra block of code that snuck into SciPy `1.5.3`
draft backports here:
https://github.com/scipy/scipy/pull/12906#discussion_r499325826

* the above change also resolves the `pycodestyle` spacing-related
CI failure observed in the PR

* fix a `np.prod()` casting issue with bleeding edge Cython versions
based on feedback in gh-12334 and local testing

* pin `numba` to version `0.47` in Travis CI based on backport PR
discussion:
https://github.com/scipy/scipy/pull/12906#issuecomment-706090217
* sparse package still pulls in a bleeding edge version of `numba`
when that is done though, so pin `sparse` to an older version as well

* remove 3 spurious lines from `THANKS.txt` that snuck in during
merge conflict resolution:
https://github.com/scipy/scipy/pull/12906#discussion_r499324982

+++ scipy/spatial/ckdtree.pyx
@@ -745,7 +745,7 @@ cdef class cKDTree:
             raise ValueError("x must consist of vectors of length %d but "
                              "has shape %s" % (int(self.m), xshape))
 
-        n = <np.intp_t> np.prod(xshape[:-1])
+        n = <np.intp_t> np.prod(xshape[:-1], dtype=np.intp)
         xx = np.ascontiguousarray(x, dtype=np.float64).reshape(n, self.m)
 
         if p < 1:
@@ -914,7 +914,7 @@ cdef class cKDTree:
         xndim = len(xshape)
 
         # allocate an array of std::vector<npy_intp>
-        n = np.prod(retshape)
+        n = np.prod(retshape, dtype=np.intp)
 
         if return_length:
             result = np.empty(retshape, dtype=np.intp)
